Manually cast to ignore warning

+++ src/main.cpp
@@ -1012,9 +1012,10 @@ int actual_main(int argc, char** argv) {
             rend.dpi_scale = dpi / dpi_default;
     }
 
-    SDL_Window* window = SDL_CreateWindow(
-        "tesh", SDL_WINDOWPOS_UNDEFINED, SDL_WINDOWPOS_UNDEFINED, 800 * rend.dpi_scale,
-        800 * rend.dpi_scale, SDL_WINDOW_SHOWN | SDL_WINDOW_RESIZABLE | SDL_WINDOW_ALLOW_HIGHDPI);
+    SDL_Window* window =
+        SDL_CreateWindow("tesh", SDL_WINDOWPOS_UNDEFINED, SDL_WINDOWPOS_UNDEFINED,
+                         (int)(800 * rend.dpi_scale), (int)(800 * rend.dpi_scale),
+                         SDL_WINDOW_SHOWN | SDL_WINDOW_RESIZABLE | SDL_WINDOW_ALLOW_HIGHDPI);
     if (!window) {
         fprintf(stderr, "SDL_CreateWindow failed: %s\n", SDL_GetError());
         return 1;
